miscellaneous formatting and re-add checksum verification to jbd2 replay code

+++ common/include/kernel/libk_decls.h
@@ -28,7 +28,7 @@ addr_t translate_user_pointer(addr_t ptr);
 uint64_t sys_time(uint64_t* tm_target);
 paging_table get_kernel_cr3();
 uint32_t crc32c_x86_3way(uint32_t, const uint8_t*, size_t);
-uint16_t crc16_calc(const void* data, size_t len, uint16_t seed = uint16_t(0));
+uint16_t crc16_calc(const void* data, size_t len, uint16_t seed = 0);
 #ifdef __cplusplus
 }
 void kfx_save();
